requirements.txt
# Edge SLM Benchmark Framework - Dependencies
# Python 3.10+ required

# === CORE ===
openai>=1.0.0               # Client API LM Studio (compat OpenAI)
httpx>=0.25.0               # HTTP client async
pyyaml>=6.0                 # Configuration files
python-dotenv>=1.0.0        # Environment variables

# === DATA & ML ===
datasets>=2.14.0            # Hugging Face datasets
pandas>=2.0.0               # Data analysis
numpy>=1.24.0               # Numerical computing
scikit-learn>=1.3.0         # Metrics (F1, accuracy, etc.)
scipy>=1.11.0               # Statistical tests (t-test, Wilcoxon)

# === LM EVALUATION HARNESS ===
lm-eval>=0.4.0              # EleutherAI evaluation harness

# === MONITORING ===
psutil>=5.9.0               # System monitoring (RAM, CPU)
tqdm>=4.66.0                # Progress bars

# === REPORTING ===
matplotlib>=3.7.0           # Visualizations
seaborn>=0.12.0             # Statistical plots
tabulate>=0.9.0             # Table formatting

# === CODE EVALUATION ===
# For HumanEval/MBPP execution
restrictedpython>=6.0       # Safe code execution (optional)

# === UTILITIES ===
jsonlines>=4.0.0            # JSONL logging
rich>=13.0.0                # Rich console output
typer>=0.9.0                # CLI framework

# === DEVELOPMENT (optional) ===
# pytest>=7.4.0             # Testing
# black>=23.0.0             # Formatting
# ruff>=0.1.0               # Linting


